Refactor LTX unit tests around shared runtime
