+++ tests/diffusion/models/ltx2/test_ltx2_pipeline.py
@@ -1,16 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
-"""Unit tests for LTX-2.3 pipeline integration.
-
-These tests verify:
-- Pipeline is properly registered in the diffusion registry
-- Post-process function is registered
-- Cache-DiT enablers are registered
-- Pipeline shares the LTX family base without inheriting LTX2-specific behavior
-- Vocoder sample rate detection logic
-- Re-export module works correctly
-"""
+"""Unit tests for the shared LTX pipeline runtime and public contracts."""
 
 import json
 import os
@@ -20,22 +11,33 @@ from types import SimpleNamespace
 import pytest
 import torch
 
+from vllm_omni.diffusion.data import DiffusionOutput
+from vllm_omni.diffusion.models.ltx2.ltx2_components import (
+    LTX2_COMPONENT_PROFILE,
+    LTX23_COMPONENT_PROFILE,
+)
+from vllm_omni.diffusion.models.ltx2.ltx2_conditioning import LTXI2VConditioningMixin
+from vllm_omni.diffusion.models.ltx2.ltx2_denoise import LTXDenoiseExecutor, LTXPhaseResult
+from vllm_omni.diffusion.models.ltx2.ltx2_guidance import (
+    LTX_LEGACY_VELOCITY_GUIDANCE,
+    LTX_OFFICIAL_X0_GUIDANCE,
+)
+from vllm_omni.diffusion.models.ltx2.ltx2_latents import LTXAVState
+from vllm_omni.diffusion.models.ltx2.ltx2_pipeline_base import LTXPipelineBase
+from vllm_omni.diffusion.models.ltx2.ltx2_recipes import LTX2_ONE_STAGE_RECIPE, LTX23_ONE_STAGE_RECIPE
+from vllm_omni.diffusion.models.ltx2.ltx2_request import LTXRequestInputs
+from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import (
+    LTX2ImageToVideoPipeline,
+    LTX2Pipeline,
+    LTX23ImageToVideoPipeline,
+    LTX23Pipeline,
+)
+from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_two_stage import (
+    LTX2ImageToVideoTwoStagesPipeline,
+    LTX2TwoStagesPipeline,
+)
+
 pytestmark = [pytest.mark.core_model, pytest.mark.cpu]
-
-
-def _make_ltx23_pipeline(sequence_parallel_size: int = 1):
-    from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23Pipeline
-
-    pipeline = object.__new__(LTX23Pipeline)
-    torch.nn.Module.__init__(pipeline)
-    pipeline.audio_vae_temporal_compression_ratio = 4
-    pipeline.audio_vae_mel_compression_ratio = 4
-    pipeline.od_config = SimpleNamespace(parallel_config=SimpleNamespace(sequence_parallel_size=sequence_parallel_size))
-    pipeline.audio_vae = SimpleNamespace(
-        latents_mean=torch.tensor(0.0),
-        latents_std=torch.tensor(1.0),
-    )
-    return pipeline
 
 
 def _make_ltx23_request_pipe(cls):
@@ -71,6 +73,183 @@ def _resolve_request_inputs_for_test(pipe, req):
         output_type="np",
         max_sequence_length=None,
     )
+
+
+def test_ltx_versions_share_runtime_without_cross_version_inheritance():
+    assert issubclass(LTX2Pipeline, LTXPipelineBase)
+    assert issubclass(LTX23Pipeline, LTXPipelineBase)
+    assert not issubclass(LTX23Pipeline, LTX2Pipeline)
+    assert LTX2Pipeline._pack_latents is LTX23Pipeline._pack_latents
+    assert LTX2Pipeline.component_profile is LTX2_COMPONENT_PROFILE
+    assert LTX23Pipeline.component_profile is LTX23_COMPONENT_PROFILE
+    assert LTX2Pipeline.one_stage_recipe is LTX2_ONE_STAGE_RECIPE
+    assert LTX23Pipeline.one_stage_recipe is LTX23_ONE_STAGE_RECIPE
+
+
+def test_ltx_one_stage_variants_share_forward_template():
+    assert LTX2Pipeline._forward_impl is LTXPipelineBase._forward_impl
+    assert LTX2ImageToVideoPipeline._forward_impl is LTXPipelineBase._forward_impl
+    assert LTX23Pipeline._forward_impl is LTXPipelineBase._forward_impl
+    assert LTX23ImageToVideoPipeline._forward_impl is LTXPipelineBase._forward_impl
+    assert LTX2Pipeline.forward is LTX23Pipeline.forward
+    assert LTX2ImageToVideoPipeline.forward is LTX23ImageToVideoPipeline.forward
+
+
+def test_ltx_versions_share_request_prompt_and_step_templates():
+    shared_methods = (
+        "_get_gemma_prompt_embeds",
+        "encode_prompt",
+        "check_inputs",
+        "_resolve_request_inputs",
+        "_prepare_prompt_context",
+        "_denoise_step",
+    )
+    for method_name in shared_methods:
+        base_method = getattr(LTXPipelineBase, method_name)
+        assert getattr(LTX2Pipeline, method_name) is base_method
+        assert getattr(LTX23Pipeline, method_name) is base_method
+
+
+def test_ltx_versions_select_guidance_without_overriding_control_flow():
+    assert LTX2Pipeline.guidance_strategy is LTX_LEGACY_VELOCITY_GUIDANCE
+    assert LTX23Pipeline.guidance_strategy is LTX_OFFICIAL_X0_GUIDANCE
+    assert LTX2Pipeline._predict_noise_for_step is LTXPipelineBase._predict_noise_for_step
+    assert LTX23Pipeline._predict_noise_for_step is LTXPipelineBase._predict_noise_for_step
+    assert LTX2Pipeline.combine_cfg_noise is LTXPipelineBase.combine_cfg_noise
+    assert LTX23Pipeline.combine_cfg_noise is LTXPipelineBase.combine_cfg_noise
+
+
+def test_ltx2_two_stage_variants_share_stage_orchestration():
+    assert issubclass(LTX2ImageToVideoTwoStagesPipeline, LTX2TwoStagesPipeline)
+    assert LTX2ImageToVideoTwoStagesPipeline.one_stage_pipeline_cls is LTX2ImageToVideoPipeline
+    assert LTX2ImageToVideoTwoStagesPipeline._run_two_stage is LTX2TwoStagesPipeline._run_two_stage
+
+
+def test_ltx_variants_share_denoise_loop_and_i2v_conditioning():
+    assert "_denoise_loop" not in LTX2Pipeline.__dict__
+    assert "_denoise_loop" not in LTX23Pipeline.__dict__
+    assert issubclass(LTX2ImageToVideoPipeline, LTXI2VConditioningMixin)
+    assert issubclass(LTX23ImageToVideoPipeline, LTXI2VConditioningMixin)
+    assert LTX2ImageToVideoPipeline.prepare_latents is LTXI2VConditioningMixin.prepare_latents
+    assert LTX23ImageToVideoPipeline.prepare_latents is LTXI2VConditioningMixin.prepare_latents
+    assert LTX2Pipeline.prepare_latents is LTXPipelineBase.prepare_latents
+    assert LTX23Pipeline.prepare_latents is LTXPipelineBase.prepare_latents
+    assert LTX2Pipeline.prepare_audio_latents is LTXPipelineBase.prepare_audio_latents
+    assert LTX23Pipeline.prepare_audio_latents is LTXPipelineBase.prepare_audio_latents
+    assert LTX2Pipeline._decode_and_split is LTXPipelineBase._decode_and_split
+    assert LTX23Pipeline._decode_and_split is LTXPipelineBase._decode_and_split
+
+
+def test_denoise_executor_owns_progress_interrupt_and_current_timestep():
+    updates = []
+
+    class Progress:
+        def __enter__(self):
+            return self
+
+        def __exit__(self, exc_type, exc, traceback):
+            return False
+
+        def update(self):
+            updates.append(True)
+
+    pipeline = SimpleNamespace(
+        _current_timestep=None,
+        interrupt=False,
+        progress_bar=lambda total: Progress(),
+    )
+    seen: list[tuple[int, float, float]] = []
+    timesteps = torch.tensor([3.0, 2.0, 1.0])
+    initial_state = LTXAVState(video=torch.tensor(0.0), audio=torch.tensor(10.0))
+
+    def step(index, timestep, state):
+        seen.append((index, timestep.item(), pipeline._current_timestep.item()))
+        pipeline.interrupt = True
+        return LTXAVState(video=state.video + 1, audio=state.audio + 1)
+
+    state = LTXDenoiseExecutor.run(pipeline, initial_state, timesteps, step)
+
+    assert seen == [(0, 3.0, 3.0)]
+    assert updates == [True]
+    torch.testing.assert_close(state.video, torch.tensor(1.0))
+    torch.testing.assert_close(state.audio, torch.tensor(11.0))
+
+
+def test_ltx2_two_stage_reuses_prompt_context_between_phases():
+    request_inputs = LTXRequestInputs(
+        prompt="prompt",
+        negative_prompt="negative",
+        height=32,
+        width=32,
+        num_frames=1,
+        frame_rate=24.0,
+        num_inference_steps=4,
+        guidance_scale=4.0,
+        guidance_rescale=0.0,
+        num_videos_per_prompt=1,
+        generator=None,
+        latents=None,
+        audio_latents=None,
+        prompt_embeds=None,
+        negative_prompt_embeds=None,
+        prompt_attention_mask=None,
+        negative_prompt_attention_mask=None,
+        decode_timestep=0.0,
+        decode_noise_scale=None,
+        output_type="np",
+        max_sequence_length=16,
+    )
+    prompt_context = object()
+    phase_calls = []
+
+    class FakePipe(torch.nn.Module):
+        def _resolve_request_inputs(self, req, **kwargs):
+            return request_inputs
+
+        def _run_denoise_phase(self, req, inputs, *, prompt_context=None, **kwargs):
+            phase_calls.append((inputs, prompt_context))
+            if len(phase_calls) == 1:
+                assert prompt_context is None
+                context = prompt_context_sentinel
+                video = torch.tensor([1.0])
+                audio = torch.tensor([2.0])
+            else:
+                assert prompt_context is prompt_context_sentinel
+                torch.testing.assert_close(inputs.latents, torch.tensor([11.0]))
+                torch.testing.assert_close(inputs.audio_latents, torch.tensor([2.0]))
+                assert inputs.guidance_scale == 1.0
+                assert inputs.num_inference_steps == 3
+                context = prompt_context
+                video = torch.tensor([3.0])
+                audio = torch.tensor([4.0])
+            return LTXPhaseResult(
+                forward_context=SimpleNamespace(prompt_context=context),
+                video=video,
+                audio=audio,
+            )
+
+        def _decode_and_split(self, forward_context, video, audio):
+            return DiffusionOutput(output=(video, audio))
+
+    class FakeUpsampler(torch.nn.Module):
+        def forward(self, *, latents, output_type, return_dict):
+            assert output_type == "latent"
+            assert not return_dict
+            return (latents + 10,)
+
+    prompt_context_sentinel = prompt_context
+    pipeline = object.__new__(LTX2TwoStagesPipeline)
+    torch.nn.Module.__init__(pipeline)
+    pipeline.distilled = True
+    pipeline.pipe = FakePipe()
+    pipeline.upsample_pipe = FakeUpsampler()
+
+    output = pipeline.forward(SimpleNamespace())
+
+    assert len(phase_calls) == 2
+    assert phase_calls[1][1] is prompt_context_sentinel
+    torch.testing.assert_close(output.output[0], torch.tensor([3.0]))
+    torch.testing.assert_close(output.output[1], torch.tensor([4.0]))
 
 
 class TestLTX23RequestParsing:
@@ -310,32 +489,7 @@ class TestLTX23ForwardStages:
         }
 
 
-class TestPipelineInheritance:
-    """Verify sibling public pipelines meet at the shared family base."""
-
-    def test_ltx23_pipeline_does_not_inherit_from_ltx2(self):
-        """LTX23Pipeline must not inherit LTX2-specific numerical behavior."""
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX2Pipeline, LTX23Pipeline
-
-        assert not issubclass(LTX23Pipeline, LTX2Pipeline), (
-            "LTX23Pipeline should inherit the shared family base, not LTX2Pipeline"
-        )
-
-    def test_ltx23_pipeline_is_nn_module(self):
-        """LTX23Pipeline must be an nn.Module."""
-        import torch.nn as nn
-
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23Pipeline
-
-        assert issubclass(LTX23Pipeline, nn.Module)
-
-    def test_ltx23_pipeline_has_progress_bar(self):
-        """LTX23Pipeline must mix in ProgressBarMixin."""
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23Pipeline
-        from vllm_omni.diffusion.models.progress_bar import ProgressBarMixin
-
-        assert issubclass(LTX23Pipeline, ProgressBarMixin)
-
+class TestPipelineComponents:
     def test_ltx23_pipeline_declares_offload_components(self):
         """LTX23Pipeline must expose LTX-2.3-specific modules to offload discovery."""
         from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23Pipeline
@@ -360,13 +514,6 @@ class TestPipelineInheritance:
         assert modules.encoder_names == ["text_encoder", "connectors"]
         assert modules.resident_names == ["vocoder"]
         assert len(modules.vaes) == 2
-
-    def test_ltx23_pipeline_has_diffusion_pipeline_profiler_mixin(self):
-        """LTX23Pipeline must support lightweight stage timing."""
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23Pipeline
-        from vllm_omni.diffusion.profiler.diffusion_pipeline_profiler import DiffusionPipelineProfilerMixin
-
-        assert issubclass(LTX23Pipeline, DiffusionPipelineProfilerMixin)
 
 
 class TestLTX23DecodeConditioning:
@@ -406,17 +553,6 @@ class TestLTX23DecodeConditioning:
 
 class TestRegistryIntegration:
     """Verify all LTX-2.3 pipeline variants are registered."""
-
-    def test_pipeline_models_registered(self):
-        """LTX-2.3 pipeline variants must be in _DIFFUSION_MODELS."""
-        from vllm_omni.diffusion.registry import _DIFFUSION_MODELS
-
-        expected = [
-            "LTX23Pipeline",
-            "LTX23ImageToVideoPipeline",
-        ]
-        for name in expected:
-            assert name in _DIFFUSION_MODELS, f"{name} not found in _DIFFUSION_MODELS"
 
     def test_pipeline_module_paths(self):
         """Registry entries must point to the correct modules."""
@@ -542,101 +678,3 @@ class TestPostProcessFunction:
         assert "video" in result
         assert "audio" in result
         assert "audio_sample_rate" not in result
-
-
-class TestInitExports:
-    """Test that __init__.py exports all LTX-2.3 classes."""
-
-    def test_all_ltx23_classes_exported(self):
-        """All LTX-2.3 pipeline classes must be in the ltx2 package __all__."""
-        from vllm_omni.diffusion.models import ltx2
-
-        expected_classes = [
-            "LTX23Pipeline",
-            "LTX23ImageToVideoPipeline",
-        ]
-        for name in expected_classes:
-            assert hasattr(ltx2, name), f"{name} not exported from ltx2 package"
-            assert name in ltx2.__all__, f"{name} not in ltx2.__all__"
-
-
-class TestAudioLatentSPPadding:
-    def test_prepare_audio_latents_pads_generated_dummy_length_for_sp(self):
-        pipeline = _make_ltx23_pipeline(sequence_parallel_size=2)
-
-        latents, original_num_frames, padded_num_frames = pipeline.prepare_audio_latents(
-            batch_size=1,
-            num_channels_latents=8,
-            num_mel_bins=64,
-            audio_latent_length=1,
-            dtype=torch.float32,
-            device=torch.device("cpu"),
-        )
-
-        assert original_num_frames == 1
-        assert padded_num_frames == 2
-        assert latents.shape == (1, 2, 128)
-
-    def test_prepare_audio_latents_pads_provided_packed_sequence_dim_for_sp(self):
-        pipeline = _make_ltx23_pipeline(sequence_parallel_size=4)
-        latents = torch.arange(40, dtype=torch.float32).view(1, 10, 4)
-
-        padded, original_num_frames, padded_num_frames = pipeline.prepare_audio_latents(
-            batch_size=1,
-            num_channels_latents=2,
-            num_mel_bins=8,
-            audio_latent_length=10,
-            dtype=torch.float32,
-            device=torch.device("cpu"),
-            latents=latents,
-        )
-
-        assert original_num_frames == 10
-        assert padded_num_frames == 12
-        assert padded.shape == (1, 12, 4)
-        torch.testing.assert_close(padded[:, :10], latents)
-        torch.testing.assert_close(padded[:, 10:], torch.zeros(1, 2, 4))
-
-    def test_prepare_audio_latents_accepts_already_padded_4d_latents_for_sp(self):
-        pipeline = _make_ltx23_pipeline(sequence_parallel_size=4)
-        latents = torch.arange(96, dtype=torch.float32).view(1, 2, 12, 4)
-
-        audio_latent_length = pipeline._resolve_audio_latent_length(10, latents)
-        padded, original_num_frames, padded_num_frames = pipeline.prepare_audio_latents(
-            batch_size=1,
-            num_channels_latents=2,
-            num_mel_bins=16,
-            audio_latent_length=audio_latent_length,
-            dtype=torch.float32,
-            device=torch.device("cpu"),
-            latents=latents,
-        )
-
-        assert audio_latent_length == 10
-        assert original_num_frames == 10
-        assert padded_num_frames == 12
-        assert padded.shape == (1, 12, 8)
-        torch.testing.assert_close(padded, pipeline._pack_audio_latents(latents))
-
-    def test_resolve_audio_latent_length_preserves_legacy_4d_shape_inference(self):
-        pipeline = _make_ltx23_pipeline(sequence_parallel_size=4)
-        latents = torch.zeros(1, 2, 13, 4)
-
-        audio_latent_length = pipeline._resolve_audio_latent_length(10, latents)
-
-        assert audio_latent_length == 13
-
-    def test_prepare_audio_latents_rejects_incompatible_provided_length(self):
-        pipeline = _make_ltx23_pipeline(sequence_parallel_size=4)
-        latents = torch.zeros(1, 11, 4)
-
-        with pytest.raises(ValueError, match="incompatible audio frame count"):
-            pipeline.prepare_audio_latents(
-                batch_size=1,
-                num_channels_latents=2,
-                num_mel_bins=8,
-                audio_latent_length=10,
-                dtype=torch.float32,
-                device=torch.device("cpu"),
-                latents=latents,
-            )
