+++ tests/diffusion/models/ltx2/test_ltx2_vae.py
